+++ fixtures/pages-functions-with-routes-app/functions/bye.ts
@@ -1,0 +1,3 @@
+export async function onRequest() {
+	return new Response("[/functions/bye]: Adieu 🙁");
+}