+++ fixtures/pages-functions-with-routes-app/functions/index.ts
@@ -1,0 +1,3 @@
+export async function onRequest() {
+	return new Response("ROOT");
+}