+++ fixtures/pages-workerjs-with-routes-app/public/_worker.js
@@ -1,20 +1,39 @@
 export default {
 	async fetch(request, env) {
 		const url = new URL(request.url);
-		if (url.pathname.startsWith("/greeting/hello")) {
-			return new Response("Bonjour le monde!");
+
+		if (url.pathname === "/") {
+			return new Response("ROOT");
 		}
 
-		if (url.pathname.startsWith("/greeting/goodbye")) {
-			return new Response("A plus tard alligator 👋");
+		if (url.pathname === "/party") {
+			return new Response(
+				"[/party]: Oops! Tous les alligators sont allés à la fête 🎉"
+			);
 		}
 
-		if (url.pathname.startsWith("/party")) {
-			return new Response("Oops! Tous les alligators sont allés à la fête 🎉");
+		if (url.pathname === "/party-disco") {
+			return new Response("[/party-disco]: Tout le monde à la discothèque 🪩");
 		}
 
-		if (url.pathname.startsWith("/date")) {
-			return new Response(new Date().toISOString());
+		if (url.pathname === "/date") {
+			return new Response(`[/date]: ${new Date().toISOString()}`);
+		}
+
+		if (url.pathname === "/greeting") {
+			return new Response("[/greeting]: Bonjour à tous!");
+		}
+
+		if (url.pathname === "/greeting/hello") {
+			return new Response("[/greeting/hello]: Bonjour le monde!");
+		}
+
+		if (url.pathname === "/greeting/bye") {
+			return new Response("[/greeting/bye]: A plus tard alligator 👋");
+		}
+
+		if (url.pathname === "/greetings") {
+			return new Response("[/greetings]: Bonjour alligators!");
 		}
 
 		return env.ASSETS.fetch(request);