+++ packages/wrangler/src/__tests__/helpers/msw/index.ts
@@ -1,9 +1,11 @@
 import { setupServer } from "msw/node";
+import { mswSuccessDeployments } from "./handlers/deployments";
 import { mswSuccessNamespacesHandlers } from "./handlers/namespaces";
 import { mswSuccessOauthHandlers } from "./handlers/oauth";
 import { mswSuccessR2handlers } from "./handlers/r2";
-import { default as mswScriptHandlers } from "./handlers/script";
+import { default as mswSucessScriptHandlers } from "./handlers/script";
 import { mswSuccessUserHandlers } from "./handlers/user";
+import { default as mswZoneHandlers } from "./handlers/zones";
 export const msw = setupServer();
 
 export {
@@ -11,5 +13,7 @@ export {
 	mswSuccessR2handlers,
 	mswSuccessOauthHandlers,
 	mswSuccessNamespacesHandlers,
-	mswScriptHandlers,
+	mswSucessScriptHandlers,
+	mswZoneHandlers,
+	mswSuccessDeployments,
 };