+++ packages/wrangler/src/__tests__/index.test.ts
@@ -34,7 +34,7 @@ describe("wrangler", () => {
 			  wrangler init [name]         📥 Create a wrangler.toml configuration file
 			  wrangler dev [script]        👂 Start a local server for developing your worker
 			  wrangler publish [script]    🆙 Publish your Worker to Cloudflare.
-			  wrangler delete [script]     🗑 Delete your Worker from Cloudflare.
+			  wrangler delete [script]     🗑  Delete your Worker from Cloudflare.
 			  wrangler tail [worker]       🦚 Starts a log tailing session for a published Worker.
 			  wrangler secret              🤫 Generate a secret that can be referenced in a Worker
 			  wrangler secret:bulk <json>  🗄️  Bulk upload secrets for a Worker
@@ -50,6 +50,7 @@ describe("wrangler", () => {
 			  wrangler login               🔓 Login to Cloudflare
 			  wrangler logout              🚪 Logout from Cloudflare
 			  wrangler whoami              🕵️  Retrieve your user info and test your auth config
+			  wrangler types               📝 Generate types from bindings & module rules in config
 
 			Flags:
 			  -c, --config   Path to .toml configuration file  [string]
@@ -77,7 +78,7 @@ describe("wrangler", () => {
 			  wrangler init [name]         📥 Create a wrangler.toml configuration file
 			  wrangler dev [script]        👂 Start a local server for developing your worker
 			  wrangler publish [script]    🆙 Publish your Worker to Cloudflare.
-			  wrangler delete [script]     🗑 Delete your Worker from Cloudflare.
+			  wrangler delete [script]     🗑  Delete your Worker from Cloudflare.
 			  wrangler tail [worker]       🦚 Starts a log tailing session for a published Worker.
 			  wrangler secret              🤫 Generate a secret that can be referenced in a Worker
 			  wrangler secret:bulk <json>  🗄️  Bulk upload secrets for a Worker
@@ -93,6 +94,7 @@ describe("wrangler", () => {
 			  wrangler login               🔓 Login to Cloudflare
 			  wrangler logout              🚪 Logout from Cloudflare
 			  wrangler whoami              🕵️  Retrieve your user info and test your auth config
+			  wrangler types               📝 Generate types from bindings & module rules in config
 
 			Flags:
 			  -c, --config   Path to .toml configuration file  [string]