+++ packages/wrangler/templates/pages-dev-pipeline.ts
@@ -2,30 +2,23 @@
 import worker from "__ENTRY_POINT__";
 // @ts-ignore entry point will get replaced
 export * from "__ENTRY_POINT__";
+import { isRoutingRuleMatch } from "./pages-dev-util";
 
-const transformToRegex = (filter: string) => {
-	return filter.replace("*", ".*");
-};
-
-const routes = {
-	// @ts-ignore routes are injected
-	include: __ROUTES__.include.map(transformToRegex),
-	// @ts-ignore routes are injected
-	exclude: __ROUTES__.exclude.map(transformToRegex) || [],
-};
+// @ts-ignore routes are injected
+const routes = __ROUTES__;
 
 export default {
 	fetch(request, env, context) {
 		const { pathname } = new URL(request.url);
 
 		for (const exclude of routes.exclude) {
-			if (pathname.match(exclude)) {
+			if (isRoutingRuleMatch(pathname, exclude)) {
 				return env.ASSETS.fetch(request);
 			}
 		}
 
 		for (const include of routes.include) {
-			if (pathname.match(include)) {
+			if (isRoutingRuleMatch(pathname, include)) {
 				return worker.fetch(request, env, context);
 			}
 		}